requirements.txt
fastapi
uvicorn
langchain
langgraph
openai
faiss-cpu
sentence-transformers
streamlit
python-dotenv
requests
pandas
numpy
tqdm